requirements.txt
Babel==2.6.0
decorator==4.3.0
Django==2.1.2
django-phonenumber-field==2.0.1
myserver==0.0.5
mysql-connector-python==8.0.12
mysqlclient==1.3.13
phonenumbers==8.9.16
pkg-resources==0.0.0
protobuf==3.6.1
pytz==2018.5
six==1.11.0
validators==0.12.2
